chore(node/util/types): Change directory hierarchy
