+++ node/_tools/config.json
@@ -36,6 +36,7 @@
       "test-util-inspect-namespace.js",
       "test-util-inspect-proxy.js",
       "test-util-inspect.js",
+      "test-util-types.js",
       "test-util.js"
     ]
   },
