+++ node/internal/stream_base_commons.ts
@@ -28,7 +28,7 @@ import {
   streamBaseState,
   WriteWrap,
 } from "../internal_binding/stream_wrap.ts";
-import { isUint8Array } from "../_util/_util_types.ts";
+import { isUint8Array } from "./util/types.ts";
 import { errnoException } from "../_errors.ts";
 import { getTimerDuration, kTimeout, setUnrefTimeout } from "../_timers.ts";
 import { validateCallback } from "../_validators.ts";
