--- node/util/types.ts
@@ -1,48 +1,0 @@
-import * as types from "../_util/_util_types.ts";
-
-export default types;
-
-export const {
-  // isExternal,
-  isDate,
-  isArgumentsObject,
-  isBigIntObject,
-  isBooleanObject,
-  isNumberObject,
-  isStringObject,
-  isSymbolObject,
-  isNativeError,
-  isRegExp,
-  isAsyncFunction,
-  isGeneratorFunction,
-  isGeneratorObject,
-  isPromise,
-  isMap,
-  isSet,
-  isMapIterator,
-  isSetIterator,
-  isWeakMap,
-  isWeakSet,
-  isArrayBuffer,
-  isDataView,
-  isSharedArrayBuffer,
-  // isProxy,
-  isModuleNamespaceObject,
-  isAnyArrayBuffer,
-  isBoxedPrimitive,
-  isArrayBufferView,
-  isTypedArray,
-  isUint8Array,
-  isUint8ClampedArray,
-  isUint16Array,
-  isUint32Array,
-  isInt8Array,
-  isInt16Array,
-  isInt32Array,
-  isFloat32Array,
-  isFloat64Array,
-  isBigInt64Array,
-  isBigUint64Array,
-  // isKeyObject,
-  // isCryptoKey
-} = types;
